Introduced support of modules
Here a bit tricky code. Let me start from use-case.

Let image that I made a library that should be compiled to JS code which
checking in runtime where it is running: on DOM, NodeJS, etc.

Some features inside nodeJS are available only via requirements.

Right now it is possibly to get access to it via `js.Dynamic.global`.

But it isn't possible to get access via typed like `@JSImport` because
it fails on linking with error like:
```
Unsupported input: List(CommonJSModule([some path]/main.js))
```

So, here a trivial changes that allows to mix
`withModuleKind(ModuleKind.CommonJSModule)` with this `jsEnv`.

I also added a test to proove that it works, and explain how it works.